revert mechanical changes
- Revert creating an apiproxy.go file and save that for a follow up to avoid creating large diffs unnecessarily.
